Refactor matrix I/O operations: separate HDF5, CSV, binary, and conversion functionalities into distinct modules. Implement save and load functions for HDF5 format, including error handling and metadata attributes. Add binary format saving functionality. Enhance CSV saving capabilities for matrices with IDs and similarity pairs. Introduce conversion functions for CSV, FASTA, and JSON formats, with placeholders for future implementation.
