+++ rs_jaccard/src/matrix_io/convert.rs
@@ -1,0 +1,44 @@
+use anyhow::Result;
+use std::fs::File;
+use std::io::{BufWriter, Write};
+use std::path::Path;
+
+/// Convert from CSV format to coordinates and bounds files
+pub fn from_csv(input: &Path, coords_out: &Path, bounds_out: &Path) -> Result<()> {
+    let file = File::open(input)?;
+    let mut reader = csv::Reader::from_reader(file);
+    
+    let coords_file = File::create(coords_out)?;
+    let mut coords_writer = BufWriter::new(coords_file);
+    
+    let bounds_file = File::create(bounds_out)?;
+    let mut bounds_writer = BufWriter::new(bounds_file);
+    
+    let mut current_pos = 0;
+    writeln!(bounds_writer, "{}", current_pos)?; // Start with 0
+    
+    for result in reader.records() {
+        let record = result?;
+        for field in record.iter() {
+            if !field.trim().is_empty() {
+                writeln!(coords_writer, "{}", field.trim())?;
+                current_pos += 1;
+            }
+        }
+        writeln!(bounds_writer, "{}", current_pos)?;
+    }
+    
+    Ok(())
+}
+
+/// Convert from FASTA format to coordinates and bounds files
+pub fn from_fasta(_input: &Path, _coords_out: &Path, _bounds_out: &Path) -> Result<()> {
+    // Implement FASTA to k-mer conversion
+    todo!("FASTA conversion not implemented yet")
+}
+
+/// Convert from JSON format to coordinates and bounds files  
+pub fn from_json(_input: &Path, _coords_out: &Path, _bounds_out: &Path) -> Result<()> {
+    // Implement JSON to k-mer conversion
+    todo!("JSON conversion not implemented yet")
+}
